getHotelRooms

+++ controllers/room.js
@@ -1,5 +1,6 @@
 import Rooms from "../models/rooms.js";
 import Hotel from '../models/hotels.js';
+import { createError } from "../utils/error.js";
 
 
 
@@ -27,11 +28,14 @@ export const addRoom = async (req, res, next) => {
     const room = new Rooms(req.body);
     try{
         const savedRoom = await room.save();
-        await Hotel.findByIdAndUpdate(
+        const updatedHotel = await Hotel.findByIdAndUpdate(
              req.params.hotelId,
              {$push: {rooms: savedRoom._id}});
-        res.status(200).setHeader('Content-Type', 'application/json');
-        res.json(savedRoom);
+       if(updatedHotel==null){
+        throw createError(404, "Hotel not found!");
+       }
+       res.status(200).setHeader('Content-Type', 'application/json');
+       res.json(savedRoom);
     }catch(err){
         next(err);
     }
@@ -49,8 +53,11 @@ export const updateRoom = async (req, res, next) => {
 
 export const deleteRoom = async (req, res, next) => {
     try{
+        const deletedHotel = await Hotel.findByIdAndUpdate(req.params.hotelId, {$pull: {rooms: req.params.id}});
+        if(deletedHotel==null){
+            throw createError(404, "Hotel not found!")
+        }
         await Rooms.findByIdAndDelete(req.params.id);
-        await Hotel.findByIdAndUpdate(req.params.hotelId, {$pull: {rooms: req.params.id}});
         res.status(200).setHeader('Content-Type', 'application/json');
         res.json({message: "Room has been deleted"});
     }catch(err){
